fix(php): use ptrdiff_t for readpipe capacity math

+++ php.c
@@ -301,7 +301,7 @@ char *php_readpipe(int php_process, char *command) {
 			bptr = result_string;
 
 			while (1) {
-				int avail = RESULTS_BUFFER - 1 - (bptr - result_string);
+				ptrdiff_t avail = RESULTS_BUFFER - 1 - (bptr - result_string);
 
 				if (avail <= 0) {
 					SPINE_LOG(("ERROR: SS[%i] The Script Server result was longer than the acceptable range", php_process));
